+++ lib/src/server.dart
@@ -89,9 +89,11 @@ abstract class StreamServer {
   /** The port. Default: 8080.
    */
   int port;
-  /** The host. Default: "127.0.0.1".
+  /** The host. Default: InternetAddress.ANY_IP_V4 (i.e., "0.0.0.0").
+   * In other words, it causes Stream server to listen all adapters
+   * IP addresses using IPv4.
    */
-  String host;
+  InternetAddress host;
 
   /** The timeout, in seconds, for sessions of this server.
    * Default: 1200 (unit: seconds)